Moved Ollama class from __init__.py to ollama.py; linted code
